:recycle: refactor: Rename ProgramSelector to ProgramWithGatesSelector
